SR-IOV migration: Reattach SR-IOV devices when aborted
When migrating a VM with SR-IOV host-devices is aborted
due to failure or is canceled by the user, the VM will
end up with no SR-IOV devices connected until the next reboot.

The reason for that is the fact that we detach SR-IOV
host-devices just before migration starts due to Libvirt
won't allow migration VM’s with connected VFIO devices
(e.g: SR-IOV VF).

To overcome this, SR-IOV host-devices will be reattached to
The target VM when a migration fails.

Currently when migration is aborted due to client cancel request
it will be handled like a migration failure on `asyncMigrate`
gorutine [1] [2].
Thus re-connecting the devices at that point covers both cases
where a migration is aborted due to failure or cancel request.

In order to test this, we use stress tool, lowering migration bandwidth
and increase the VM memory request in order to slow down the migration
following live migration test-suite example.
It will increase the time the migration is on phase running
and will enable us to fail or cancel it.

[1] https://github.com/kubevirt/kubevirt/blob/36a9eb7659fc56c1b150081c818afc12544efbef/pkg/virt-launcher/virtwrap/live-migration-source.go#L893
[2] https://github.com/kubevirt/kubevirt/blob/36a9eb7659fc56c1b150081c818afc12544efbef/pkg/virt-launcher/virtwrap/live-migration-source.go#L917-L923

Signed-off-by: Or Mergi <ormergi@redhat.com>