sriov migration: Reattach sriov devices when aborted
When migrating a VM with SRIOV host-devices is aborted
due to failure or is canceled by the user, the VM will
end up with no SRIOV devices connected until the next reboot.

The reason for that is the fact that we detach SRIOV
host-devices just before migration starts due to Libvirt
won't allow migration VM’s with connected VFIO devices
(e.g: SRIOV VF).

To overcome this, SRIOV host-devices will be reattached to
The target VM when a migration fails.

In order to test this, we use stress tool in order to
slow down the migration following live migration test suite
example.
It will increase the time the migration is on phase running
and will enable us to fail or cancel it.

Signed-off-by: Or Mergi <ormergi@redhat.com>