virt-launcher: Shutdown interfaces DHCP servers when unplugged
When an interface is hot-plugged, depending on its IPAM config,
a DHCP server instance is created to provide its IP address.
Once unplugged, the corresponding DHCP server instance should be
disposed.

Both DHCPv4 and DHCPv6 implementations converts the given connection to
raw-connection, which make its to "lose" its close(), when called it
does nothing and the connection still open.
Thus decorating the connection with context (using the net.ListenConfig API)
wont work, when the context is canceled, the connection wont close and
the DHCP server still run in the background.

To overcome this, we pass context to the DHCP server on creation and
start a Gorutine that wait for the context to canceled.
When it does, close the underlying connection which will cause the server
to shutdown.

Since the current implementing of PodNIC is stateless, each instance is
re-generated appon usage, we cannot not use it (i.e: having each instance
keep its own context or any stop-channel alternative).
virt-launcher's domain-manager is extended to keep track of each interface
DHCP server instance context as it live long as the pod is, and enable
controlling each DHCP server instance knob.

Signed-off-by: Or Mergi <ormergi@redhat.com>