virt-controller, vm: Avoid unplugging interfaces on older running VMs
When marking a network interface for removal, the VM controller is
patching the VMI object to reflect the same on the running VM.

However, for older VMs where its pod interfaces use ordered naming, the
unplug operation is not supported.

Therefore, the VM controller now checks against the pod network
annotation about the naming of the pod interfaces and if detected as
ordered, it will not patch the VMI.

The change assures VMI objects which do not support interface hot-unplug
will not be marked as such in the first place.

Signed-off-by: Edward Haas <edwardh@redhat.com>