virt-handler, VMI iface status: Keep detached secondary ifaces status
Currently an unplugged iface reporting is flaky due to a race between
virt-handler and virt-controller, where virt-handler set the VMI iface
status according to the domain which dosent include detached interfaces.

On the other hande virt-controller see the subject external interface
stil exist (according to the CNI annotation) and add it to the VMI status.

The end result is unpluged interfaces status is flaky, where it keeps get
deleted and added.

To fix this, keep detached interfaces status whos  external iface exist
in pod (have info-source mutlus-status) and also exist in the spec.

Signed-off-by: Or Mergi <ormergi@redhat.com>